requirements.txt
duckduckgo-search>=6.0.0
google-generativeai
requests
beautifulsoup4